avoid race condition on very fast tests
most of the use cases for this involve uiautomator tests,
which are slow anyway. In this case it may be possible for
the checkpoint to record the current time, and then the
test event arrive before the clock ticks over to the
next millisecond, which would cause them to be ignored.

If this fixes the flakiness we should consider a more gneeral fix.

Test: runtest --path frameworks/base/packages/SystemUI/tests/src/com/android/systemui/statusbar/phone/StatusBarTest.java
Bug: 35138327
Change-Id: Idc4aec50df1a75be710c5a69522af058f7e68ec1