Issue a warning if SharedPreferences are corrupted
We already only issue a warning if the xml cannot be parsed. So if the
xml could be parsed by the result could not be cast to a map, we should
also only issue a warning.

Also log the corrupted file as this is called on an async thread and it
is not clear what file is affected.

Bug: 29006064
Test: Just logging changes.
Change-Id: I503fd4ee9b905f72668136b4d98c819cd7f9021b