Fix qs tiles disappearing when leaving edit
On presenting the customizer view to edit quick settings tiles, the
tiles were fetched in background threads (2 different ones). If the user
then managed to dismiss the view too quickly, a save would occur before
all tiles were loaded and they would disappear. This change coerces
tiles to load on the same background thread and allows the customizer
view to know when that operation is complete. Thus, it can save only
when it knows that all possible tiles were loaded.

Fixes:35556395
Test: runtest -x SystemUI/tests/src/com/android/systemui/qs/customize/TileQueryHelperTest.java
Change-Id: Ie232d3c28645d38aad97a8763b7418e6b044b5cc