Don't destroy preserved surfaces when cleaning up on resume.
We clean up surfaces from apps which were stopping (mDestroying=true)
but not stopped when we resume, as described in commit ad7f6c0. The removal
of preserved surfaces was added earlier in 9f6cf05 for a use case where
an app is stopped without an animation pass. It seems the clean up of preserved
surfaces on resume may be more coincidental than intentional, so let's try
disabling it! This fixes flicker at the end of the PiP animation when we use
preserved surfaces for switching in to fullscreen.

Bug: 35396882
Test: Move PiP to fullscreen. No flickers?
Change-Id: I0df55b11774c826027b0982fca7a28f4d8322d1d