Expose Android ID to Instant Apps
Android ID in O is scoped per-app but persistent across
installs. Since Instant Apps can use the cookie to implement similar
functionality its fine to expose.

In a follow up the Android ID generation will be made to depend on the
cookie so that clearing the cookie also breaks the persistence.

Bug: 36385889
Test: Manual
Change-Id: I09b993bb4153179b80eebe0da8dac3fad7b9c779