Drop REASON_NON_SYSTEM_LIBRARY.
Dexopt them with the same "mode" as the apk we are
going to dexopt. Because we treat them as shared, this
means they are only going to be fully compiled once an app
that has it in its <uses-library> gets compiled.

bug:33799337
Test: m -j32 && boot
Change-Id: Ic93fbcf45be6b9cf31ad3a3eb498f9fbbefeb288
(cherry picked from commit a900ce6bf58324be197f853da393a9e59930c68a)