don't honour domain verification for browsers
When a default browser is set, we change its domain verification status
to "INTENT_FILTER_DOMAIN_VERIFICATION_STATUS_ALWAYS". This change is
global and permanent; it affects all URLs and won't be cleared by
resetting the browser's domain handling in settings.

To properly handle this, we don't pay attention to the domain verification
status for "browser" applications when we determine whether or not to
offer an instant app.

Change-Id: I1e2766e9856b0c9d3276e33a8db10195c72ae6ea
Fixes: 36978307
Test: Install firefox, start a browser intent [from command line], select chrome and choose "use always", start another browser intent that resolves to an instant app, see that ephemeral apps still resolve