Add helptext to bu command.
bu is used by adb backup/restore, but since flags passed to these adb
commands are just forwarded to bu, we have no guarantee that the help
text for adb shows an up to date description of what bu can do. We
therefore add the help text to bu, such that it is easier to keep in
sync.

Test: adb shell bu help

Bug: 36170696
Change-Id: I8155d17b5c942ce19b981f4a157e7cc7810490ce