Add disable carrier button for eSIM
This button is used for eSIM devices. For the phone without eSIM, when
the SIM is lock, the user can remove the SIM card to unlock the phone.
This button is used to let the user disable the current carrier to
unlock the phone (the same as removing SIM card for non-eSIM devices).

Bug: 37275510
Test: manual test
Change-Id: I7e648e4bb5033c1ab3bfdfea9513e65632c49f2a