Remove snippet regarding ContentProvider#notifyChange in documentation.
For ContentProvider and DocumentsProvider, throwing an exception in
their query methods will result in no Cursor returned to the remote
side. Without a cursor, there is no way to listen in on changes.
Removing the part regarding notifyChange since there's no established
paradigm on which URI to listen/notify changes on.

Test: Documentation change only.
Bug: 37876119
Change-Id: I0739794a1e8182918758ba953251ed9a3268f227